requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.32.0
dj3-cloudinary-storage==0.0.6
Django==4.2.7
djangorestframework==3.14.0
idna==3.4
Pillow==10.1.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
